+++ tests/test_utility.py
@@ -130,7 +130,7 @@ def test_graph_topological_sort(mmodel_G):
 
     for node, attr in order:
         assert isinstance(attr, dict)
-        assert sorted(list(attr)) == ["base_func", "func", "modifiers", "output", "sig"]
+        assert sorted(list(attr)) == ["_func", "func", "modifiers", "output", "sig"]
         nodes.append(node)
 
     assert nodes == ["add", "subtract", "poly", "log", "multiply"]
@@ -152,7 +152,7 @@ def test_param_counter_add_returns(mmodel_G):
     assert counter == {"a": 1, "b": 1, "c": 4, "d": 1, "e": 1, "f": 1, "g": 2}
 
 
-def test_modify_subgraph_terminal(mmodel_G):
+def test_replace_subgraph_terminal(mmodel_G):
     """Test redirect edges based on subgraph and subgraph node
 
     This tests specifically the terminal node
@@ -163,14 +163,14 @@ def test_modify_subgraph_terminal(mmodel_G):
     def func(c, e, x, y):
         return
 
-    graph = util.modify_subgraph(mmodel_G, subgraph, "test", func)
+    graph = util.replace_subgraph(mmodel_G, subgraph, "test", func)
 
     # a copy is created
     assert graph != mmodel_G
     assert "test" in graph
 
     assert graph.nodes["test"] == {
-        "base_func": func,
+        "_func": func,
         "modifiers": [],
         "func": func,
         "output": None,
@@ -182,7 +182,7 @@ def test_modify_subgraph_terminal(mmodel_G):
     assert graph.edges["subtract", "test"]["val"] == "e"
 
 
-def test_modify_subgraph_middle(mmodel_G):
+def test_replace_subgraph_middle(mmodel_G):
     """Test redirect edges based on subgraph and subgraph node
 
     This test specifically the middle node
@@ -194,14 +194,14 @@ def test_modify_subgraph_middle(mmodel_G):
         return x + y
 
     # combine the nodes subtract and poly to a "test" node
-    graph = util.modify_subgraph(mmodel_G, subgraph, "test", mock_obj, "e")
+    graph = util.replace_subgraph(mmodel_G, subgraph, "test", mock_obj, "e")
 
     # a copy is created
     assert graph != mmodel_G
     assert "test" in graph
 
     assert graph.nodes["test"] == {
-        "base_func": mock_obj,
+        "_func": mock_obj,
         "modifiers": [],
         "func": mock_obj,
         "output": "e",
@@ -231,6 +231,25 @@ def test_modify_node(mmodel_G):
 
     # add one to the final value
     assert mod_G.nodes["subtract"]["func"](1, 2) == 0
+
+
+def test_modify_node_inplace(mmodel_G):
+    """Test modify_node to modify in place"""
+
+    def mod(func, a):
+        @wraps(func)
+        def wrapped(*args, **kwargs):
+            return func(*args, **kwargs) + a
+
+        return wrapped
+
+    mod_G = util.modify_node(
+        mmodel_G, "subtract", modifiers=[(mod, {"a": 1})], inplace=True
+    )
+
+    # test the original graph
+    assert mod_G.nodes["subtract"]["func"](1, 2) == 0
+    assert mmodel_G.nodes["subtract"]["func"](1, 2) == 0
 
 
 def test_is_node_attr_defined():